+++ packages/snaps-utils/src/manifest/validation.ts
@@ -34,6 +34,7 @@ import { isDerivationPathEqual } from '../array';
 import { CronjobSpecificationArrayStruct } from '../cronjob';
 import { SIP_6_MAGIC_VALUE, STATE_ENCRYPTION_MAGIC_VALUE } from '../entropy';
 import { KeyringOriginsStruct, RpcOriginsStruct } from '../json-rpc';
+import { KeyringCapabilitiesStruct } from '../keyring';
 import { SnapIdStruct } from '../snaps';
 import { mergeStructs, type InferMatching } from '../structs';
 import { NameStruct, NpmSnapFileNames, uri } from '../types';
@@ -216,7 +217,11 @@ export const PermissionsStruct: Describe<InitialPermissions> = type({
   ),
   'endowment:ethereum-provider': optional(EmptyObjectStruct),
   'endowment:keyring': optional(
-    mergeStructs(HandlerCaveatsStruct, KeyringOriginsStruct),
+    mergeStructs(
+      HandlerCaveatsStruct,
+      KeyringOriginsStruct,
+      KeyringCapabilitiesStruct,
+    ),
   ),
   'endowment:protocol': optional(
     mergeStructs(